moved some helper functions
