+++ packages/periphery/test/integration/TriggerOrders/Optimism.test.ts
@@ -18,14 +18,14 @@ import {
 import { impersonate } from '../../../../common/testutil'
 import { parse6decimal } from '../../../../common/testutil/types'
 import { transferCollateral } from '../../helpers/marketHelpers'
-import { createMarketETH, deployProtocol, deployPythOracleFactory } from '../../helpers/setupHelpers'
+import { deployPythOracleFactory } from '../../helpers/oracleHelpers'
+import { createMarketETH, deployProtocol } from '../../helpers/setupHelpers'
 import { RunManagerTests } from './Manager.test'
 import { FixtureVars } from './setupTypes'
 import {
   CHAINLINK_ETH_USD_FEED,
   DSU_ADDRESS,
   DSU_RESERVE,
-  mockGasInfo,
   PYTH_ADDRESS,
   USDC_ADDRESS,
   USDC_HOLDER,
